requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2024.8.30
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
frozenlist==1.5.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
magic-filter==1.0.12
more-itertools==8.10.0
multidict==6.1.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==21.3
pexpect==4.8.0
propcache==0.2.0
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydantic==2.7.4
pydantic_core==2.18.4
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-magic==0.4.24
PyYAML==5.4.1
redis==5.2.1
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.7.2
ssh-import-id==5.11
systemd-python==234
Telethon==1.38.1
texttable==1.6.4
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
websocket-client==1.2.3
xkit==0.0.0
yarl==1.17.2
zipp==1.0.0
zope.interface==5.4.0
